Move enum_traits.h to common for cross-module enum use
