+++ tools/codegen/templates/enum.h.j2
@@ -9,7 +9,7 @@
 #include <magic_enum/magic_enum.hpp>
 {% endif %}
 
-#include "data/enums/enum_traits.h"
+#include "common/enum_traits.h"
 
 namespace xi
 {
